refactor: modernize C++ idioms across src/plot and src/plot3d
- Add override to ~120 derived class destructors (2D + 3D modules)

- Add override to 7 virtual function overrides in 3D module

  (fixes Label::draw() and ColorLegend::draw() missing virtual keyword)

- Replace Q_DISABLE_COPY with = delete in 18 files (19 classes)

- Convert 12 old-style SIGNAL/SLOT connect() to new-style function pointers

- Replace 7 C-style casts with static_cast or remove redundant casts

- Replace QWidget* parent = 0 with nullptr in QwtLegendLabel

- Add plan_modern_cpp_phase2.md for follow-up improvements

+++ src/plot/qwt_date_scale_engine.cpp
@@ -627,7 +627,7 @@ static QwtScaleDiv qwtDivideToMonths(
         }
         else if ( minStepSize > 0.0 )
         {
-            const int numMinorSteps = qRound( stepSize / (double) minStepSize );
+            const int numMinorSteps = qRound( stepSize / minStepSize );
 
             for ( int i = 1; i < numMinorSteps; i++ )
             {
